+++ CHANGELOG.md
@@ -2,14 +2,18 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
-- FIX `requestIdlePromise()` must run in a queue.
-- ADD Export ReplicationOptions type [#4606](https://github.com/pubkey/rxdb/pull/4606)
-- ADD `mode` option to remote storage
+
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->
 
 <!-- RELEASE BELOW -->
+
+### 14.6.1 (6 April 2023)
+
+- FIX `requestIdlePromise()` must run in a queue.
+- ADD Export ReplicationOptions type [#4606](https://github.com/pubkey/rxdb/pull/4606)
+- ADD `mode` option to remote storage
 
 ### 14.6.0 (31 March 2023)